remove .gitenv from .profile and move it to bash_aliases
There is a bug execution .gitenv from .profile because
xession is sourcing with sh not bash, so we get syntax errors
in ~/.xsession-errors.   Not sure why this started happening
now on trusty latest patches.  Non the less moving .gitenv
to .bash_aliases fixes this problem.

Also added a method for stoping git-login